Code refactoring and improvements
Fixes V2 Bug Tracker Ahmedsaed/gdsc-cert#11
