+++ pages/404.js
@@ -1,8 +1,5 @@
-import React, { useEffect, useState } from "react";
-import firebase from "firebase/app";
+import React from "react";
 import "firebase/firestore";
-import Head from "next/head";
-import Cert from "../components/cert";
 
 export default function CertPage() {
     return (
