+++ pages/_app.js
@@ -3,7 +3,7 @@ import firebase from "firebase/app";
 import "firebase/analytics";
 import Head from "next/head";
 import { useEffect } from "react";
-import Logout from "../components/logout";
+import Logout from "../components/Logout";
 
 if (!firebase.apps.length) {
     firebase.initializeApp({
@@ -26,6 +26,7 @@ function MyApp({ Component, pageProps }) {
         <>
             <Head>
                 <title>GDSC Certificates</title>
+                <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css" />
                 <link
                     rel="apple-touch-icon"
                     sizes="180x180"
